amended classes to implement serializable allowing data to pass through activities, correct button now returning random item

+++ app/src/test/java/com/example/user/thegame/GameTest.java
@@ -36,7 +36,7 @@ public class GameTest {
 
     @Test
     public void testLength(){
-        assertEquals(3, game.getLength());
+        assertEquals((Integer)3, game.getLength());
     }
 
     @Test
